+++ lib/EasyPost/EasyPostObject.php
@@ -2,6 +2,7 @@
 
 namespace EasyPost;
 
+use EasyPost\Util\InternalUtil;
 use EasyPost\Util\Util;
 
 class EasyPostObject implements \ArrayAccess, \Iterator
@@ -143,7 +144,7 @@ class EasyPostObject implements \ArrayAccess, \Iterator
     public function convertEach($client, $values)
     {
         foreach ($values as $k => $v) {
-            $this->_values[$k] = Util::convertToEasyPostObject($client, $v, $this, $k);
+            $this->_values[$k] = InternalUtil::convertToEasyPostObject($client, $v, $this, $k);
         }
     }