+++ lib/EasyPost/Service/OrderService.php
@@ -4,7 +4,7 @@ namespace EasyPost\Service;
 
 use EasyPost\Http\Requestor;
 use EasyPost\Rate;
-use EasyPost\Util\Util;
+use EasyPost\Util\InternalUtil;
 
 /**
  * Order service containing all the logic to make API calls.
@@ -54,7 +54,7 @@ class OrderService extends BaseService
         $url = $this->instanceUrl(self::$modelClass, $id) . '/rates';
         $response = $requestor->request('get', $url, $params);
 
-        return Util::convertToEasyPostObject($this->client, $response);
+        return InternalUtil::convertToEasyPostObject($this->client, $response);
     }
 
     /**
@@ -78,6 +78,6 @@ class OrderService extends BaseService
 
         $response = $requestor->request('post', $url, $params);
 
-        return Util::convertToEasyPostObject($this->client, $response);
+        return InternalUtil::convertToEasyPostObject($this->client, $response);
     }
 }